* minisip w32 port: Compilation error and warning fixes for contactdb    code.
git-svn-id: svn://svn.minisip.org/minisip/trunk@1833 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/minisip/contactdb/PhoneBook.cxx
@@ -105,7 +105,7 @@ string PhoneBookPerson::getName(){
 }
 
 void PhoneBookPerson::addEntry( MRef<ContactEntry *> entry ){
-        entry->personIndex = entries.size();
+        entry->personIndex = (uint32_t)entries.size();
 	entries.push_back( entry );
 }
